Split the Core instructions
