+++ src/main/scala/parsley/internal/instructions/CoreInstrs.scala
@@ -6,7 +6,6 @@ import parsley.internal.UnsafeOption
 import parsley.internal.deepembedding.Parsley
 
 import scala.annotation.tailrec
-import scala.language.existentials
 
 // Stack Manipulators
 private [internal] final class Push[A](x: A) extends Instr {
@@ -29,68 +28,6 @@ private [internal] object Flip extends Instr {
         ctx.exchangeAndContinue(x)
     }
     override def toString: String = "Flip"
-}
-
-// Primitives
-private [internal] class CharTok(c: Char, x: Any, _expected: UnsafeOption[String]) extends Instr {
-    val expected: String = if (_expected == null) "\"" + c + "\"" else _expected
-    override def apply(ctx: Context): Unit = {
-        if (ctx.moreInput && ctx.nextChar == c) {
-            ctx.consumeChar()
-            ctx.pushAndContinue(x)
-        }
-        else ctx.fail(expected)
-    }
-    override def toString: String = if (x == c) s"Chr($c)" else s"ChrPerform($c, $x)"
-}
-
-private [internal] final class Satisfies(f: Char => Boolean, expected: UnsafeOption[String]) extends Instr {
-    override def apply(ctx: Context): Unit = {
-        if (ctx.moreInput && f(ctx.nextChar)) ctx.pushAndContinue(ctx.consumeChar())
-        else ctx.fail(expected)
-    }
-    override def toString: String = "Sat(?)"
-}
-
-private [internal] final class StringTok private [instructions] (s: String, x: Any, _expected: UnsafeOption[String]) extends Instr {
-    private [this] val expected = if (_expected == null) "\"" + s + "\"" else _expected
-    private [this] val cs = s.toCharArray
-    private [this] val sz = cs.length
-    private [this] val adjustAtIndex = new Array[(Int => Int, Int => Int)](s.length + 1)
-    def makeAdjusters(col: Int, line: Int, tabprefix: Option[Int]): (Int => Int, Int => Int) =
-        if (line > 0) ((_: Int) => col, (x: Int) => x + line)
-        else (tabprefix match {
-            case Some(prefix) =>
-                val outer = 4 + col + prefix
-                val inner = prefix - 1
-                (x: Int) => outer + x - ((x + inner) & 3)
-            case None => (x: Int) => x + col
-        }, (x: Int) => x)
-    @tailrec def compute(cs: Array[Char], i: Int = 0, col: Int = 0, line: Int = 0)(implicit tabprefix: Option[Int] = None): Unit = {
-        adjustAtIndex(i) = makeAdjusters(col, line, tabprefix)
-        if (i < cs.length) cs(i) match {
-            case '\n' => compute(cs, i + 1, 1, line + 1)(Some(0))
-            case '\t' if tabprefix.isEmpty => compute(cs, i + 1, 0, line)(Some(col))
-            case '\t' => compute(cs, i + 1, col + 4 - ((col - 1) & 3), line)
-            case _ => compute(cs, i + 1, col + 1, line)
-        }
-    }
-    compute(cs)
-
-    @tailrec private def go(ctx: Context, i: Int, j: Int): Unit = {
-        if (j < sz && i < ctx.inputsz && ctx.input(i) == cs(j)) go(ctx, i + 1, j + 1)
-        else {
-            val (colAdjust, lineAdjust) = adjustAtIndex(j)
-            ctx.col = colAdjust(ctx.col)
-            ctx.line = lineAdjust(ctx.line)
-            ctx.offset = i
-            if (j < sz) ctx.fail(expected)
-            else ctx.pushAndContinue(x)
-        }
-    }
-
-    override def apply(ctx: Context): Unit = go(ctx, ctx.offset, 0)
-    override def toString: String = if (x.isInstanceOf[String] && (s eq x.asInstanceOf[String])) s"Str($s)" else s"StrPerform($s, $x)"
 }
 
 // Applicative Functors
@@ -131,19 +68,6 @@ private [internal] object Return extends Instr {
     override def toString: String = "Return"
 }
 
-private [internal] final class Fail(msg: String, expected: UnsafeOption[String]) extends Instr {
-    override def apply(ctx: Context): Unit = {
-        ctx.fail(expected)
-        ctx.raw ::= msg
-    }
-    override def toString: String = s"Fail($msg)"
-}
-
-private [internal] final class Unexpected(msg: String, expected: UnsafeOption[String]) extends Instr {
-    override def apply(ctx: Context): Unit = ctx.fail(expected = expected, unexpected = msg)
-    override def toString: String = s"Unexpected($msg)"
-}
-
 private [internal] final class Empty(expected: UnsafeOption[String]) extends Instr {
     override def apply(ctx: Context): Unit = {
         val strip = ctx.expected.isEmpty
@@ -168,38 +92,6 @@ private [internal] final class PushFallthrough(var label: Int) extends JumpInstr
         ctx.inc()
     }
     override def toString: String = s"PushFallthrough($label)"
-}
-
-private [internal] object Attempt extends Instr {
-    override def apply(ctx: Context): Unit = {
-        // Remove the recovery input from the stack, it isn't needed anymore
-        if (ctx.status eq Good) {
-            ctx.states = ctx.states.tail
-            ctx.handlers = ctx.handlers.tail
-            ctx.inc()
-        }
-        // Pop input off head then fail to next handler
-        else {
-            ctx.restoreState()
-            ctx.fail()
-        }
-    }
-    override def toString: String = "Attempt"
-}
-
-private [internal] object Look extends Instr {
-    override def apply(ctx: Context): Unit = {
-        if (ctx.status eq Good) {
-            ctx.restoreState()
-            ctx.handlers = ctx.handlers.tail
-            ctx.inc()
-        }
-        else {
-            ctx.states = ctx.states.tail
-            ctx.fail()
-        }
-    }
-    override def toString: String = "Look"
 }
 
 private [internal] final class InputCheck(var label: Int) extends JumpInstr {
@@ -235,31 +127,6 @@ private [internal] object Catch extends Instr {
         ctx.checkStack = ctx.checkStack.tail
     }
     override def toString: String = s"Catch"
-}
-
-// Position Extractors
-private [internal] object Line extends Instr {
-    override def apply(ctx: Context): Unit = ctx.pushAndContinue(ctx.line)
-    override def toString: String = "Line"
-}
-
-private [internal] object Col extends Instr {
-    override def apply(ctx: Context): Unit = ctx.pushAndContinue(ctx.col)
-    override def toString: String = "Col"
-}
-
-// Register-Manipulators
-private [internal] final class Get(v: Int) extends Instr {
-    override def apply(ctx: Context): Unit = ctx.pushAndContinue(ctx.regs(v))
-    override def toString: String = s"Get($v)"
-}
-
-private [internal] final class Put(v: Int) extends Instr {
-    override def apply(ctx: Context): Unit = {
-        ctx.copyOnWrite(v, ctx.stack.peekAndExchange(()))
-        ctx.inc()
-    }
-    override def toString: String = s"Put($v)"
 }
 
 // Debugging Instructions
@@ -312,13 +179,4 @@ private [internal] final class LogEnd(val name: String, break: Boolean) extends 
         if (break) doBreak(ctx)
     }
     override def toString: String = s"LogEnd($name)"
-}
-
-// Companion Objects
-private [internal] object CharTok {
-    def apply(c: Char, expected: UnsafeOption[String]): Instr = new CharTok(c, c, expected)
-}
-
-private [internal] object StringTok {
-    def apply(s: String, expected: UnsafeOption[String]): StringTok = new StringTok(s, s, expected)
 }
